+++ api-gateway/src/test/java/com/petclinic/bffapigateway/presentationlayer/BFFApiGatewayControllerFilterIntegrationTest.java
@@ -59,7 +59,7 @@ class BFFApiGatewayControllerFilterIntegrationTest {
     @MockBean private JwtTokenUtil jwtTokenUtil;
     @MockBean private Utility utility;
 
-//    @Test
+// @Test
     void testGetAllCustomers_ShouldReturnUnauthorized() {
 
         client.get()
