+++ petclinic-frontend/src/features/customers/api/addPet.ts
@@ -3,7 +3,8 @@ import axiosInstance from '@/shared/api/axiosInstance';
 import { PetRequestModel } from '@/features/customers/models/PetRequestModel.ts';
 
 export const addPet = async (
+  ownerId: string,
   pet: PetRequestModel
 ): Promise<AxiosResponse<PetRequestModel>> => {
-  return axiosInstance.post(`/pets`, pet, { useV2: false });
+  return axiosInstance.post(`owners/${ownerId}/pets`, pet, { useV2: true });
 };
