+++ petclinic-frontend/src/features/customers/api/getPet.ts
@@ -3,9 +3,13 @@ import axiosInstance from '@/shared/api/axiosInstance';
 import { PetResponseModel } from '../models/PetResponseModel.ts';
 
 export const getPet = async (
+  ownerId: string,
   petId: string
 ): Promise<AxiosResponse<PetResponseModel>> => {
-  return await axiosInstance.get<PetResponseModel>(`/pets/${petId}`, {
-    useV2: true,
-  });
+  return await axiosInstance.get<PetResponseModel>(
+    `/owners/${ownerId}/pets/${petId}`,
+    {
+      useV2: true,
+    }
+  );
 };
