+++ petclinic-frontend/src/features/customers/api/updatePet.ts
@@ -3,8 +3,13 @@ import axiosInstance from '@/shared/api/axiosInstance';
 import { PetRequestModel } from '@/features/customers/models/PetRequestModel.ts';
 
 export const updatePet = async (
+  ownerId: string,
   petId: string,
   pet: PetRequestModel
 ): Promise<AxiosResponse<void>> => {
-  return await axiosInstance.put<void>(`/pets/${petId}`, pet, { useV2: true });
+  return await axiosInstance.put<void>(
+    `/owners/${ownerId}/pets/${petId}`,
+    pet,
+    { useV2: true }
+  );
 };
